msm: pm-stats: fix the bug in pm stats refactor
Return value of msm_timer_get_sclk_time api is int64_t. Storing the
return value in integer would lead to inconsistencies in pm_stats.

CRs-fixed: 375038

+++ arch/arm/mach-msm/pm-8x60.c
@@ -590,7 +590,7 @@ static void msm_pm_timer_exit_idle(bool timer_halted)
 
 static int64_t msm_pm_timer_enter_suspend(int64_t *period)
 {
-	int time = 0;
+	int64_t time = 0;
 
 	if (msm_pm_use_qtimer)
 		return sched_clock();
